Add files via upload
An audio watermarking system based on my own modified version of DSS (Direct Spread Spectrum) this would watermark audio with unique hash stored on an RDS database which linked it personal information to allow the tracking of unauthorized/leaked music. The goal was to make sure the watermark was undetectable by human ear and not compromise audio quality.